pyproject.toml
[project]
name = "ws-conan1-chs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "freetype-py>=2.5.1",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
]
